% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbData.R
\docType{data}
\name{data-Chatterjee}
\alias{data-Chatterjee}
\alias{Chatterjee}
\title{Chatterjee, Hadi and Price Data}
\format{The format is a dataframe of 40 observations and 7 variables.}
\source{
The data come from Chatterjee, Hadi and Price (2000).
}
\usage{
data(Chatterjee)
}
\description{
Chatterjee, Hadi and Price Data
}
\examples{
data(Chatterjee)
}
\references{
Chatterjee, S., Hadi, A. S., and Price, B. (2000). \emph{Regression Analysis by Example}. Hoboken: Wiley.
}
\keyword{datasets}
