% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbData.R
\docType{data}
\name{data-OlymWomen}
\alias{data-OlymWomen}
\alias{OlymWomen}
\title{1984 Los Angeles Olympic records data of track events for women}
\format{The format is a dataframe of 55 observations and 8 variables.}
\source{
The data come from Khattree and Naik (2000, pp. 511-512).
}
\usage{
data(OlymWomen)
}
\description{
Data are time records from 1984 Olympic track events for women from 55 countries: 100-meter, 200-meter, 400-meter, 800-meter, 1500-meter, 3000-meter, and Marathon. The corresponding variables are named as m100, m200, m400, m800, m1500, m3000, and marathon. The time measurements are recorded in seconds.
}
\examples{
data(OlymWomen)
}
\references{
Khattree, R. and Naik, D. (2000). \emph{Multivariate Data Reduction and Discrimination with SAS® Software}. Cary, NC: SAS Institute Inc.
}
\keyword{datasets}
