% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbData.R
\docType{data}
\name{data-SwissHead}
\alias{data-SwissHead}
\alias{SwissHead}
\title{Measurements of Heads of Swiss Soldiers}
\format{The format is a dataframe of 200 observations and 6 variables.}
\source{
The data come from Flurry and Riedwyl (1988).
}
\usage{
data(SwissHead)
}
\description{
Data are measurements, in millimeters, of the heads of 200 Swiss soldiers.
}
\examples{
data(SwissHead)
}
\references{
Flurry, B. and Riedwyl, H. (1988). \emph{Multivariate Statistics: A Practical Approach}. London: Chapman and Hall.
}
\keyword{datasets}
